requirements.txt
selenium==3.141.0
pytest==6.2.4
pytest-azurepipelines==0.8.0
pytest-cov==2.12.1
grpcio==1.38.1
pytest-grpc==0.8.0
grpcio-tools==1.38.1